refactor wheel encoder driver hw test
