+++ packages/wheel_encoder/include/hardware_test_wheel_encoder/__init__.py
@@ -1,0 +1,1 @@
+from .hardware_test_wheel_encoder import HWTestWheelEncoder
